+++ src/test/java/dev/solar/lotto/domain/LottoTest.java
@@ -1,4 +1,4 @@
-package dev.solar.lotto;
+package dev.solar.lotto.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.withPrecision;
